+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ### Added
 
 - Added Trunk Support to improve Code Quality
+- Added context menu (right click) option to delete report files
 
 ### Changed